refactor(lot): rename dc_lot to lot
