+++ src/context.rs
@@ -6,13 +6,13 @@ use crate::contact::*;
 use crate::dc_job::*;
 use crate::dc_jobthread::*;
 use crate::dc_loginparam::*;
-use crate::dc_lot::Lot;
 use crate::dc_move::*;
 use crate::dc_msg::*;
 use crate::dc_receive_imf::*;
 use crate::dc_tools::*;
 use crate::imap::*;
 use crate::key::*;
+use crate::lot::Lot;
 use crate::param::Params;
 use crate::smtp::*;
 use crate::sql::Sql;
