+++ src/dc_msg.rs
@@ -9,8 +9,8 @@ use crate::constants::*;
 use crate::contact::*;
 use crate::context::*;
 use crate::dc_job::*;
-use crate::dc_lot::Lot;
 use crate::dc_tools::*;
+use crate::lot::Lot;
 use crate::param::*;
 use crate::pgp::*;
 use crate::sql;
