+++ src/dc_qr.rs
@@ -4,10 +4,10 @@ use crate::chat;
 use crate::constants::Blocked;
 use crate::contact::*;
 use crate::context::Context;
-use crate::dc_lot::Lot;
 use crate::dc_strencode::*;
 use crate::dc_tools::*;
 use crate::key::*;
+use crate::lot::Lot;
 use crate::param::*;
 use crate::peerstate::*;
 use crate::types::*;
