+++ src/lib.rs
@@ -32,6 +32,7 @@ pub mod context;
 mod imap;
 pub mod key;
 pub mod keyring;
+pub mod lot;
 pub mod oauth2;
 mod param;
 pub mod peerstate;
@@ -51,7 +52,6 @@ pub mod dc_job;
 mod dc_jobthread;
 pub mod dc_location;
 mod dc_loginparam;
-pub mod dc_lot;
 mod dc_mimefactory;
 pub mod dc_mimeparser;
 mod dc_move;
